Add mask, VGA mode, NTSC and fixed-scanline mode
Mask
----

CRT shadow mask uses a simple subpixel implementation with 2-pixel,
2x2-pixel, and 4-pixel patterns.

At 1440p, the 2-pixel pattern is good for trinitron or dotmask (with
a 1-pixel stagger), and the 2x2 pattern without a stagger is a pretty
convincing slot mask.

4-pixel is probably good at 4k, but I don't have anything to test it on.

The min and max transparency of the mask effect can be changed
individually, to simulate a bit of bloom and allow brighter whites.  To
avoid harsh transitions in the effect, the transparency scaling is based
on a box-filtered source pixel value.

Interlacing/VGA Modes
---------------------

Interlacing can now be disabled, and interlacing trigger resolution is
configurable (but 320 lines is probably good for pretty much
everything). Above the trigger resolution, alternate scanlines will be
drawn each frame.

VGA mode is provided as a second "interlacing" mode using the same
trigger resolution. Below the trigger resolution, every line will be
drawn using two half-size scanlines. This is needed becuase real VGA
hardware would output the 320x200 resolution mode using 400 horizontal
lines, but DOSBox just outputs a 320x200 image instead of doubling it to
320x400.

Fixed-Scanline Mode
-------------------

You can now provide a "reference" number of scanlines, roughly the
number of horizontal lines that the monitor was designed to display. The
size of the output scanline will be scaled according to the square root
of the difference between the actual and reference number of scanlines.

If we assume a fixed output power (and no letterboxing), this should
roughly match the behavior of a real CRT when it switches resolutions.
You can also give it a size slightly smaller than the output resolution,
and it will blend lines together vertically without (much) artifacting.

NTSC Simlulation
----------------

The NTSC shader from guest-advanced is integrated into a separate
crt-ntsc.slangp preset. In this preset this shader just provides the
scanline effect; the color blending is performed by guest-advanced.

In comparison to the YIQ color mode in this shader, guest-advanced also
simulates intermodulation artifacts like color bleed, fringing, and dot
crawl.

--

All of this together means you can simulate not only a CRT TV, but also
any 1990s PC monitor--even the crazy stuff you could attach to a Sharp
x68000 (which supported 256x256 double-strike, 512x512 interlaced,
512x512 progressive, and 1024x1024). And you only need one preset per
system.

`crt-simple.slangp` is the same as `crt.slangp` but without the
full-screen blur pass (I'm not a fan).