Add Famicom/NES NTSC color generation
This is a new shader based loosely on nes-color-decoder.slang and the
nesdev wiki, replacing the first pass of crt-guest-advanced-ntsc.

This shader is intended for use with "RAW" palette output mode in
Nestopia (and probably other emulators). It emulates the Famicom video
output to convert the PPU image into an NTSC signal, which is then
converted to an image using the remaining passes of
crt-guest-advanced-ntsc, before having the scanlines drawn by this
shader.

This provides more accurate colors than can be provided otherwise, since
about half of the visible color generated by the Famicom is created with
NTSC artifact colors.

You can also tweak the output palette by changing the crosstalk
(artifacting and fringing) values in the shader parameters.

The American NES and Japanese Famicom apparently output exactly the same
signal, but American NTSC specifies a higher black level than NTSC-J
(NTSC-J black is "below black" in NTSC). If it matters to you, I've
added an option to tune to NTSC black level. By default it uses NTSC-J.

`ntsc-pass1-famicom.slang` is fully compatible with
crt-guest-advanced-ntsc.